requirements.txt
tushare>=1.4.0
openai>=1.40.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
